requirements.txt
mysql-connector-python==2.1.3
pysolr==3.5.0
requests==2.11.0
